+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.5.0
+- Less noisy output: Only show number of unpushed commits in output if there are more than 0.
+
 # 1.4.0
 - The list of commits included in a mob branch is now truncated to a maximum of 5 entries to prevent the need for scrolling up in order to see the latest included changes.
 - Show more informative error message when `mob <cmd>` is run outside of a git repository.